+++ routes/devices.js
@@ -32,7 +32,7 @@ router.get(deviceApiUrls.deviceStatus, function(req, res, next)
 	}
 	else
 	{
-		res.status(404).send();
+		res.status(404).send("Page Does Not Exist");
 	}
 });
 
@@ -85,21 +85,63 @@ router.post(deviceApiUrls.deviceType, function(req, res, next)
 
 router.get(deviceApiUrls.deviceQuery, function(req, res, next)
 {
-	// Todo
-	res.send("Read Device");
+	var prepDeviceID = deviceParams.readPage(req.params.deviceId);
+	
+	deviceParams.retrieveDevice(prepDeviceID, function(retrievedData)
+	{
+		if (retrievedData.outcome > 0)
+		{
+			res.send(retrievedData.deviceInfo);
+		}
+		else if (retrievedData.outcome === 0)
+		{
+			res.status(400).send(retrievedData.messageText);
+		}
+		else
+		{
+			res.status(404).send(retrievedData.messageText);
+		}
+	});
 });
 
 router.put(deviceApiUrls.deviceQuery, function(req, res, next)
 {
-	// Todo
-	res.send("Modify Device");
+	var updateResultObject = {};
+	
+	rioIndex.modRemoteIoDevice(req.body, function (updateExistingErr, updatedID)
+	{
+		if (updateExistingErr !== null)
+		{
+			res.status(400).send(updateExistingErr.message);
+		}
+		else
+		{
+			updateResultObject["success"] = true;
+			updateResultObject["id"] = updatedID;
+			res.send(updateResultObject);
+		}
+	});
 });
 
 
 router.delete(deviceApiUrls.deviceQuery, function(req, res, next)
 {
-	// Todo
-	res.send("Delete Device");
+	var deleteProp = req.headers["delete-permanently"];
+	var delStatus = (deleteProp === "true");
+	var deleteResultObject = {};
+	
+	rioIndex.delRemoteIoDevice(req.params.deviceId, delStatus, function (deleteExistingErr)
+	{
+		if (deleteExistingErr !== null)
+		{
+			res.status(400).send(deleteExistingErr.message);
+		}
+		else
+		{
+			deleteResultObject["success"] = true;
+			res.send(deleteResultObject);
+		}
+	});
 });